4.21 Complete

+++ controllers/blogs.js
@@ -20,7 +20,7 @@ blogsRouter.get('/', async (request, response) => {
   blogsRouter.post('/', async (request, response, next) => {
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if(!decodedToken.id){
-      return response.status(401).json({error: 'token Invalid'})
+      return response.status(402).json({error: 'token Invalid'})
     }
 
     const blog = request.body
@@ -41,8 +41,25 @@ blogsRouter.get('/', async (request, response) => {
   })
 
   blogsRouter.delete('/:id', async (request, response, next) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!decodedToken.id){
+      response.status(402).json({error: 'token Invalid'})
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+
+      const user = await User.findById(decodedToken.id)
+      user.blogs.remove(blog._id)
+      await user.save()
+
+      response.status(204).end()
+    } else {
+      response.status(403).json({error: 'user not authorized to delete this blog'})
+    }
+
   })
 
   blogsRouter.put('/:id', async (request, response, next) => {
